+++ packages/tools/transpiler/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @dlightjs/transpiler
 
+## 1.0.0-alpha.25
+
+### Patch Changes
+
+- babel-preset-dlight@1.0.0-alpha.25
+
 ## 1.0.0-alpha.24
 
 ### Patch Changes